added a common.js for cookies
function to set 2 user cookies, it takes 2 parameters, it also sets an expiration date of 5 hours (easily adjustable) to each cookie.

function to return a cookie value. Parameter should be the cookie key, in this case they are written "userEmail" and "userId" value returning is Email or userId as a string.

function checkCookie is not finished. I was using W3schools for this.

Linked the website for citation and reference purposes