Created a login page script
Log in checks if email attempting to log in is a current user within user list, if so, log in function creates cookie for user logging in and then redirects the user to the home page. Cookie is now created, and when a user creates a new property userID is a new hidden field within the property object. this can be used to tie a unique user to their properties