added ownerID as property to property constructor
also there is a auto fill fields for creating a new property for testing purposes. I clearly commented where the testing auto fill starts and ends. This made it easier so I did not have to fill out a property for each test. I only had to click the yes and no fields.