fixed check for existing email
Users can only create a profile if the proposed email is not already in record. The program also checks only in lower case, as emails are not case sensitive. We still need to update the output if the email is already a registered user. Currently it is still tied to the testing output