profile / edit profile page should be complete
a user can view their details, click edit to update them. The program checks if the updated email is currently in the system or not. If not then all update fields are updated to the appropriate user in the users local storage data.

the page then returns to the regular display profile page