removed view properties link displaying as default minor other tweaks
common.js was having issues being used as other js files were loading first and trying to use the methods. I added another script src for common at the top and also changed the order at the bottom. This has fixed the issue for the time being, however being aware of it is helpful.

view properties link now ONLY displays if a user is logged in as an OWNER
there should be no reason for that link to display otherwise as unregistered users and renters won't have properties.