Added a property id to the properties.
Property Id saves to the local storage and pulls it to get current Id, adds a ++ to the Id then uses that Id for the property being created. The last used Id is then saved to the local storage again.

This ensures that all property Ids are unique since it is an id value tracked by the "server"