Added a view property page for single property
When a user clicks on a property div in the home page it leads to this html page + userId and propertyId added to the url.

The script then has a self invoking function that gathers both Id's from the url then uses them to find the according user and property from the local files. After finding the correct user and property it displays their data on the page.

This page will likely have a lot of issues working unless you create a user in your local storage so that an appropriate user is connected to the ID.

I will likely create dummy "user" data to pull from as well so that this function works always.

perhaps adding some if / else statements on the variable created using the .find method might be a good idea. This will make sure that the entire page does not crash if there was an issue retrieving the data