Created the myProperties client side script
includes a lot of the old functions however a lot has been updated

editIndex is now updated to hold the value of the propertyId. This facilitated adding the correct property to an update request in order to fulfill that request.

save property now has 2 different fetch requests. One if the editIndex is greater than 0, indicating that it is an update (PUT) request.
Or if the editIndex is -1 the request is to create (POST) a new property.

UserId is still parsed from cookie data.

displayProperties function currently has AN AUTOMATIC setUserCookie function to SET DUMMY COOKIE DATA. This is due to the fact that we have no log in capabilities YET.

display Properties runs a get request to the server, using the userId (COOKIE ID currently DUMMY Cookie), this allows the server to do the filtering of properties and return  ONLY the properties that belong to that specific user, and then display them.

A few changes made to the edit and remove button event listeners.

REMOVE PROPERTY CURRENTLY DOES NOT WORK